Add recursive watch events
We register a connection-specific watch callback with Irmin, which
is called with the absolute paths of paths which are created /
updated or deleted. The callback takes care of dealing with absolute
versus relative paths where

- a watch for an absolute path /foo/bar should always receive absolute
  paths in response
- a watch for a relative path foo/bar should receive paths relative
  to the connections home directory i.e. /local/domain/<domid>/foo/bar

Note this generates more watches than are strictly required: xenstored
normally only generates sub-path watches when the nodes are being created,
while this code will generate the sub-pack watches when a node is updated.
However it should be fine to generate more watches than are strictly
required.

Signed-off-by: David Scott <dave.scott@citrix.com>