`list` should return keys in the order they were created
Some clients (notably Linux netfront) enumerates VIFs in the order
of `list` results, rather than using the natural order of the VIF
defice numbers. One could argue that VIF/NIC enumeration order
shouldn't be relied upon anyway and that one should rely on something
like MAC address matching, however this confuses enough people that
it's worth keeping the old behaviour.

When we explicitly create a path `/a/b/c`, we now update the key
`/a/b.order` to contain a list with `c` on the end. The `list`
function will return these ordered keys first, and then include
the rest for completeness. We don't bother preserving the order
for keys that were implicitly created, such as `/a`.

The directory creation order unit test now passes.

Signed-off-by: David Scott <dave.scott@citrix.com>