When testing, instantiate the `Effects` functor once per empty store
The `Effects` functor creates a module which contains the transactions
hashtable. Previously we created a fresh module on every call to `rpc`
which meant that the hashtable kept forgetting our transaction ids.

This patch modifies `empty_store` to return the result of the `Effects`
functor, not simply the raw `Xirmin` module.

This fixes the broken unit-test.

Signed-off-by: David Scott <dave.scott@citrix.com>